requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.24.0
google-generativeai>=0.3.0
pydub>=0.25.1
gTTS>=2.3.1
gitpython>=3.1.30